refactor: refactor uiSchemaTransfomer
OKTA-299837
