+++ packages/@okta/i18n/src/properties/login.properties
@@ -709,10 +709,6 @@ oie.resend.link = Send again
 ## OIE Password
 oie.password = Password
 oie.password.challenge.title = Sign in using your password
-
-## OIE Password
-oie.password = Password
-oie.password.challenge.title = Sign in using your password
 oie.password.enroll.title = Set up password
 oie.password.confirmPasswordLabel = Re-enter password
 oie.password.match.error = New passwords must match
