+++ src/v2/ion/uiSchemaTransformer.js
@@ -14,103 +14,14 @@
  * Create UI Schema into remedation object base on remediation values
  */
 import { _ } from 'okta';
+import createUiSchemaForBoolean from './ion-type-handlers/boolean';
+import createUiSchemaForObject from './ion-type-handlers/object';
+import createUiSchemaForString from './ion-type-handlers/string';
 
-const ionOptionsToUiOptions = (options) => {
-  const result = {};
-  options.forEach(({ value, label }) => {
-    result[value] = label;
-  });
-  return result;
-};
-
-/**
- * Adds `factorType` metadata to each select-factor option item using factors array
- * @param {FactorOption[]} options
- * @param {Factor[]} factors
- * @returns {Object} option
- * @returns {string} option.label
- * @returns {string} option.value
- * @returns {string} option.factorType
- */
-const createFactorTypeOptions = (options, factors) => {
-  _.each(options, function (optionItem) {
-    const factorValue = optionItem.value;
-    const factor = factors.find(function (item) {
-      return (item.factorProfileId === factorValue
-        || item.factorId === factorValue);
-    }) || {};
-    optionItem.factorType = factor.factorType;
-  });
-  return options;
-};
-
-/**
- * Example of the option like
- * @param {AuthenticatorOption[]} options
- * @param {( AuthenticatorEnrollment[] || Authenticator[] )} authenticators
- */
-const createAuthenticatorOptions = (options = []) => {
-  return options.map(option => {
-    const value = option.value && option.value.form && option.value.form.value || [];
-
-    // Each authenticator option has list of ION field.
-    // Currently we are only support merely selecting one of options
-    // rather than pop up another page to collection extra data
-    // (in order to fill value for `mutable: true; value: null` fields).
-    // The only reason of such design is to simplify widget implementation
-    // but could subject to change in later releases.
-    // Thus only surface up fields that are `required: true; mutable: false`
-    // which implies it already has `value`.
-    const valueObject = value
-      .filter(v => v.required === true && v.mutable === false)
-      .reduce((init, v) => {
-        return Object.assign(init, { [v.name]: v.value });
-      }, {});
-
-    return {
-      label: option.label,
-      value: valueObject,
-      authenticatorType: option.relatesTo && option.relatesTo.type,
-    };
-  });
-};
-
-const getCheckboxUiSchema = ({ label, type, required }) => ({
-  // For Remember Me checkbox, we need the label only on the right side of it.
-  placeholder: label,
-  label: false,
-  // Separating prop type for Backbone.Model
-  // from html input type
-  modelType: type,
-  // uiSchema type is the html input type desired.
-  type: 'checkbox',
-  required: required || false,
-});
-
-const getPasswordUiSchema = () => ({
-  type: 'password',
-  params: {
-    showPasswordToggle: true,
-  },
-});
-
-const getFactorsUiSchema = ({ options }, factors) => ({
-  type: 'factorSelect',
-  options: createFactorTypeOptions(options, factors),
-});
-
-const getAuthenticatorsEnrollUiSchema = ({ options }) => {
-  return {
-    type: 'authenticatorEnrollSelect',
-    options: createAuthenticatorOptions(options),
-  };
-};
-
-const getAuthenticatorsVerifyUiSchema = ({ options }) => {
-  return {
-    type: 'authenticatorVerifySelect',
-    options: createAuthenticatorOptions(options),
-  };
+const UISchemaHandlers = {
+  string: createUiSchemaForString,
+  boolean: createUiSchemaForBoolean,
+  object: createUiSchemaForObject,
 };
 
 /**
@@ -119,7 +30,6 @@ const getAuthenticatorsVerifyUiSchema = ({ options }) => {
  * @param {IONForm} remeditationForm
  */
 const createUISchema = (transformedResp, remediationForm) => {
-  /* eslint complexity: 0, max-statements: 0, max-depth: [2, 4] */
 
   // For cases where field itself is a form, it has a formname and we are appending the formname to each field.
   // Sort of flat the structure in order to align Courage flatten Model. The flatten structure will be converted
@@ -146,82 +56,18 @@ const createUISchema = (transformedResp, remediationForm) => {
     .value();
 
   return remediationValue.map(ionFormField => {
-    const uiSchema = {
+    const uiSchemaDefaultConfig = {
       'label-top': true,
     };
-
-    if (ionFormField.type === 'string' || !ionFormField.type) {
-      uiSchema.type = 'text';
-
-      // e.g. { name: 'password', secret: true }
-      if (ionFormField.secret === true) {
-        Object.assign(uiSchema, getPasswordUiSchema());
-      }
-
-      if (Array.isArray(ionFormField.options)) {
-        if (ionFormField.name === 'factorId'
-          || ionFormField.name === 'factorProfileId') {
-          // select factor form for multiple factor enroll and multiple factor verify
-          // when factor has not been enrolled we get back factorProfileId, and once its enrolled
-          // we get back factorId
-          // e.g. { name: 'factorId' | 'factorProfileId', type: 'string', options: [ {label: 'xxx', value: 'yyy'} ]}
-          const factors = transformedResp.factors && transformedResp.factors.value || [];
-          Object.assign(uiSchema, getFactorsUiSchema(ionFormField, factors));
-        } else if (ionFormField.name.indexOf('methodType') >= 0) {
-          // e.g. { name: 'methodType', options: [ {label: 'sms'} ], type: 'string' | null }
-          uiSchema.type = 'radio';
-          // set the default value to the first value..
-          if (ionFormField.options[0] && ionFormField.options[0].value) {
-            ionFormField.value = ionFormField.options[0].value;
-          }
-        } else {
-          // default to select (dropdown). no particular reason (certainly can default to radio.)
-          // e.g. { name: 'questionKey', options: [], type: 'string' | null }
-          uiSchema.type = 'select';
-          uiSchema.wide = true;
-          uiSchema.options = ionOptionsToUiOptions(ionFormField.options);
-        }
-      }
-    }
-
-    if (ionFormField.type === 'boolean') {
-      Object.assign(uiSchema, getCheckboxUiSchema(ionFormField));
-    }
-
-    if (ionFormField.type === 'object') {
-      if (ionFormField.name === 'authenticator' && remediationForm.name === 'select-authenticator-authenticate') {
-        // similar to `factorId` but `authenticator` is a new way to model factors
-        // hence it has different structure
-        Object.assign(uiSchema, getAuthenticatorsVerifyUiSchema(ionFormField));
-      } else if (ionFormField.name === 'authenticator' && remediationForm.name === 'select-authenticator-enroll') {
-        // TODO: OKTA-302497: use different type for enrollment flow.
-        Object.assign(uiSchema, getAuthenticatorsEnrollUiSchema(ionFormField));
-      } else {
-        // e.g. { "name": "credentials", "type": "object", options: [ {value: {form: value:[]} ]
-        uiSchema.optionsUiSchemas = ionFormField.options.map(opt => {
-          return createUISchema(transformedResp, {
-            value: [
-              {
-                name: ionFormField.name,
-                value: opt.value,
-              }
-            ]
-          });
-        });
-        uiSchema.options = ionFormField.options.map((opt, index) => {
-          return { label: opt.label, value: index };
-        });
-        // assume (default to) use radio buttons to switch sub-schema.
-        uiSchema.type = 'radio';
-        uiSchema.value = '0';
-        uiSchema.name = `sub_schema_local_${ionFormField.name}`;
-      }
-    }
+    const fieldType = ionFormField.type || 'string';
+    const uiSchemaHandler = UISchemaHandlers[fieldType];
+    const uiSchemaResult = uiSchemaHandler(ionFormField, remediationForm, transformedResp, createUISchema);
 
     return Object.assign(
       {},
       ionFormField,
-      uiSchema,
+      uiSchemaDefaultConfig,
+      uiSchemaResult,
     );
   });
 };
@@ -231,10 +77,13 @@ const createUISchema = (transformedResp, remediationForm) => {
  * @param {AuthResult} transformedResp
  */
 const insertUISchema = (transformedResp) => {
-
   if (transformedResp) {
     transformedResp.remediations = transformedResp.remediations.map(obj => {
       obj.uiSchema = createUISchema(transformedResp, obj);
+
+      const logs = _.pick(obj, 'value', 'uiSchema', 'name', 'relatesTo');
+      logs.value = logs.value && logs.value.filter(v => v.name !== 'stateHandle');
+      // console.log(JSON.stringify(logs, null, 2));
       return obj;
     });
   }
@@ -242,7 +91,6 @@ const insertUISchema = (transformedResp) => {
 };
 
 module.exports = insertUISchema;
-
 
 /**
  * @typedef {Object} FactorOption
